refactor: split definition schemas into own files
