+++ integrations/teams/definitions/user.ts
@@ -1,0 +1,14 @@
+import { IntegrationDefinitionProps } from '@botpress/sdk'
+
+export const user = {
+  tags: {
+    id: {
+      title: 'ID',
+      description: 'Teams user ID',
+    },
+    email: {
+      title: 'Email',
+      description: 'Email address',
+    },
+  },
+} satisfies IntegrationDefinitionProps['user']
